Improved code structure

+++ src/string/my_strpbrk.c
@@ -9,8 +9,8 @@
 
 char *my_strpbrk(const char *haystack, const char *list)
 {
-    for (int i = 0; haystack[i] != '\0'; i++)
-        for (int j = 0; list[j] != '\0'; j++)
+    for (unsigned long i = 0; haystack[i] != '\0'; i++)
+        for (unsigned long j = 0; list[j] != '\0'; j++)
             if (haystack[i] == list[j])
                 return (char *) &haystack[i];
